@@ -1,108 +1,106 @@
-﻿using System;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
 using StardewValley;
 
-namespace GenericModConfigMenu
+namespace AudioDescription;
+
+/// <summary>The API which lets other mods add a config UI through Generic Mod Config Menu.</summary>
+public interface IGenericModConfigMenuApi
 {
-    /// <summary>The API which lets other mods add a config UI through Generic Mod Config Menu.</summary>
-    public interface IGenericModConfigMenuApi
-    {
-        /*********
-        ** Methods
-        *********/
-        /****
-        ** Must be called first
-        ****/
-        /// <summary>Register a mod whose config can be edited through the UI.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="reset">Reset the mod's config to its default values.</param>
-        /// <param name="save">Save the mod's current config to the <c>config.json</c> file.</param>
-        /// <param name="titleScreenOnly">Whether the options can only be edited from the title screen.</param>
-        /// <remarks>Each mod can only be registered once, unless it's deleted via <see cref="Unregister"/> before calling this again.</remarks>
-        void Register(IManifest mod, Action reset, Action save, bool titleScreenOnly = false);
+    /*********
+     ** Methods
+     *********/
+    /****
+     ** Must be called first
+     ****/
+    /// <summary>Register a mod whose config can be edited through the UI.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="reset">Reset the mod's config to its default values.</param>
+    /// <param name="save">Save the mod's current config to the <c>config.json</c> file.</param>
+    /// <param name="titleScreenOnly">Whether the options can only be edited from the title screen.</param>
+    /// <remarks>Each mod can only be registered once, unless it's deleted via <see cref="Unregister"/> before calling this again.</remarks>
+    void Register(IManifest mod, Action reset, Action save, bool titleScreenOnly = false);
 
 
-        /****
-        ** Basic options
-        ****/
-        /// <summary>Add a section title at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="text">The title text shown in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the title, or <c>null</c> to disable the tooltip.</param>
-        void AddSectionTitle(IManifest mod, Func<string> text, Func<string> tooltip = null);
+    /****
+     ** Basic options
+     ****/
+    /// <summary>Add a section title at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="text">The title text shown in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the title, or <c>null</c> to disable the tooltip.</param>
+    void AddSectionTitle(IManifest mod, Func<string> text, Func<string> tooltip = null);
 
-        /// <summary>Add an image at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="texture">The image texture to display.</param>
-        /// <param name="texturePixelArea">The pixel area within the texture to display, or <c>null</c> to show the entire image.</param>
-        /// <param name="scale">The zoom factor to apply to the image.</param>
-        void AddImage(IManifest mod, Func<Texture2D> texture, Rectangle? texturePixelArea = null, int scale = Game1.pixelZoom);
+    /// <summary>Add an image at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="texture">The image texture to display.</param>
+    /// <param name="texturePixelArea">The pixel area within the texture to display, or <c>null</c> to show the entire image.</param>
+    /// <param name="scale">The zoom factor to apply to the image.</param>
+    void AddImage(IManifest mod, Func<Texture2D> texture, Rectangle? texturePixelArea = null, int scale = Game1.pixelZoom);
 
-        /// <summary>Add a boolean option at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="getValue">Get the current value from the mod config.</param>
-        /// <param name="setValue">Set a new value in the mod config.</param>
-        /// <param name="name">The label text to show in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
-        /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
-        void AddBoolOption(IManifest mod, Func<bool> getValue, Action<bool> setValue, Func<string> name, Func<string> tooltip = null, string fieldId = null);
+    /// <summary>Add a boolean option at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="getValue">Get the current value from the mod config.</param>
+    /// <param name="setValue">Set a new value in the mod config.</param>
+    /// <param name="name">The label text to show in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
+    /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
+    void AddBoolOption(IManifest mod, Func<bool> getValue, Action<bool> setValue, Func<string> name, Func<string> tooltip = null, string fieldId = null);
 
-        /// <summary>Add an integer option at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="getValue">Get the current value from the mod config.</param>
-        /// <param name="setValue">Set a new value in the mod config.</param>
-        /// <param name="name">The label text to show in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
-        /// <param name="min">The minimum allowed value, or <c>null</c> to allow any.</param>
-        /// <param name="max">The maximum allowed value, or <c>null</c> to allow any.</param>
-        /// <param name="interval">The interval of values that can be selected.</param>
-        /// <param name="formatValue">Get the display text to show for a value, or <c>null</c> to show the number as-is.</param>
-        /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
-        void AddNumberOption(IManifest mod, Func<int> getValue, Action<int> setValue, Func<string> name, Func<string> tooltip = null, int? min = null, int? max = null, int? interval = null, Func<int, string> formatValue = null, string fieldId = null);
+    /// <summary>Add an integer option at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="getValue">Get the current value from the mod config.</param>
+    /// <param name="setValue">Set a new value in the mod config.</param>
+    /// <param name="name">The label text to show in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
+    /// <param name="min">The minimum allowed value, or <c>null</c> to allow any.</param>
+    /// <param name="max">The maximum allowed value, or <c>null</c> to allow any.</param>
+    /// <param name="interval">The interval of values that can be selected.</param>
+    /// <param name="formatValue">Get the display text to show for a value, or <c>null</c> to show the number as-is.</param>
+    /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
+    void AddNumberOption(IManifest mod, Func<int> getValue, Action<int> setValue, Func<string> name, Func<string> tooltip = null, int? min = null, int? max = null, int? interval = null, Func<int, string> formatValue = null, string fieldId = null);
 
-        /// <summary>Add a string option at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="getValue">Get the current value from the mod config.</param>
-        /// <param name="setValue">Set a new value in the mod config.</param>
-        /// <param name="name">The label text to show in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
-        /// <param name="allowedValues">The values that can be selected, or <c>null</c> to allow any.</param>
-        /// <param name="formatAllowedValue">Get the display text to show for a value from <paramref name="allowedValues"/>, or <c>null</c> to show the values as-is.</param>
-        /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
-        void AddTextOption(IManifest mod, Func<string> getValue, Action<string> setValue, Func<string> name, Func<string> tooltip = null, string[] allowedValues = null, Func<string, string> formatAllowedValue = null, string fieldId = null);
+    /// <summary>Add a string option at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="getValue">Get the current value from the mod config.</param>
+    /// <param name="setValue">Set a new value in the mod config.</param>
+    /// <param name="name">The label text to show in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the field, or <c>null</c> to disable the tooltip.</param>
+    /// <param name="allowedValues">The values that can be selected, or <c>null</c> to allow any.</param>
+    /// <param name="formatAllowedValue">Get the display text to show for a value from <paramref name="allowedValues"/>, or <c>null</c> to show the values as-is.</param>
+    /// <param name="fieldId">The unique field ID for use with <see cref="OnFieldChanged"/>, or <c>null</c> to auto-generate a randomized ID.</param>
+    void AddTextOption(IManifest mod, Func<string> getValue, Action<string> setValue, Func<string> name, Func<string> tooltip = null, string[] allowedValues = null, Func<string, string> formatAllowedValue = null, string fieldId = null);
 
-        /****
-        ** Multi-page management
-        ****/
-        /// <summary>Start a new page in the mod's config UI, or switch to that page if it already exists. All options registered after this will be part of that page.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="pageId">The unique page ID.</param>
-        /// <param name="pageTitle">The page title shown in its UI, or <c>null</c> to show the <paramref name="pageId"/> value.</param>
-        /// <remarks>You must also call <see cref="AddPageLink"/> to make the page accessible. This is only needed to set up a multi-page config UI. If you don't call this method, all options will be part of the mod's main config UI instead.</remarks>
-        void AddPage(IManifest mod, string pageId, Func<string> pageTitle = null);
+    /****
+     ** Multi-page management
+     ****/
+    /// <summary>Start a new page in the mod's config UI, or switch to that page if it already exists. All options registered after this will be part of that page.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="pageId">The unique page ID.</param>
+    /// <param name="pageTitle">The page title shown in its UI, or <c>null</c> to show the <paramref name="pageId"/> value.</param>
+    /// <remarks>You must also call <see cref="AddPageLink"/> to make the page accessible. This is only needed to set up a multi-page config UI. If you don't call this method, all options will be part of the mod's main config UI instead.</remarks>
+    void AddPage(IManifest mod, string pageId, Func<string> pageTitle = null);
 
-        /// <summary>Add a link to a page added via <see cref="AddPage"/> at the current position in the form.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="pageId">The unique ID of the page to open when the link is clicked.</param>
-        /// <param name="text">The link text shown in the form.</param>
-        /// <param name="tooltip">The tooltip text shown when the cursor hovers on the link, or <c>null</c> to disable the tooltip.</param>
-        void AddPageLink(IManifest mod, string pageId, Func<string> text, Func<string> tooltip = null);
+    /// <summary>Add a link to a page added via <see cref="AddPage"/> at the current position in the form.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="pageId">The unique ID of the page to open when the link is clicked.</param>
+    /// <param name="text">The link text shown in the form.</param>
+    /// <param name="tooltip">The tooltip text shown when the cursor hovers on the link, or <c>null</c> to disable the tooltip.</param>
+    void AddPageLink(IManifest mod, string pageId, Func<string> text, Func<string> tooltip = null);
 
 
-        /****
-        ** Advanced
-        ****/
+    /****
+     ** Advanced
+     ****/
 
-        /// <summary>Register a method to notify when any option registered by this mod is edited through the config UI.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        /// <param name="onChange">The method to call with the option's unique field ID and new value.</param>
-        /// <remarks>Options use a randomized ID by default; you'll likely want to specify the <c>fieldId</c> argument when adding options if you use this.</remarks>
-        void OnFieldChanged(IManifest mod, Action<string, object> onChange);
+    /// <summary>Register a method to notify when any option registered by this mod is edited through the config UI.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    /// <param name="onChange">The method to call with the option's unique field ID and new value.</param>
+    /// <remarks>Options use a randomized ID by default; you'll likely want to specify the <c>fieldId</c> argument when adding options if you use this.</remarks>
+    void OnFieldChanged(IManifest mod, Action<string, object> onChange);
 
-        /// <summary>Remove a mod from the config UI and delete all its options and pages.</summary>
-        /// <param name="mod">The mod's manifest.</param>
-        void Unregister(IManifest mod);
-    }
+    /// <summary>Remove a mod from the config UI and delete all its options and pages.</summary>
+    /// <param name="mod">The mod's manifest.</param>
+    void Unregister(IManifest mod);
 }
