@@ -1,56 +1,50 @@
-using ContentPatcher;
-using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
 using StardewModdingAPI.Enums;
 using StardewModdingAPI.Events;
 using StardewValley;
-using System;
 
-namespace ImmersiveGrandpa
+namespace ImmersiveGrandpa;
+
+public class ModEntry : Mod
 {
-    public class ModEntry : Mod
+    public override void Entry(IModHelper helper)
     {
-        public override void Entry(IModHelper helper)
-        {
-            helper.Events.GameLoop.GameLaunched += OnGameLaunch;
-            helper.Events.Specialized.LoadStageChanged += LoadStageChange;
-        }
+        helper.Events.GameLoop.GameLaunched += OnGameLaunch;
+        helper.Events.Specialized.LoadStageChanged += LoadStageChange;
+    }
 
-         private void OnGameLaunch(object sender, GameLaunchedEventArgs e)
-        {
-            var api = this.Helper.ModRegistry.GetApi<IContentPatcherAPI>("Pathoschild.ContentPatcher");
+    private void OnGameLaunch(object sender, GameLaunchedEventArgs e)
+    {
+        var api = this.Helper.ModRegistry.GetApi<IContentPatcherAPI>("Pathoschild.ContentPatcher");
             
-            api.RegisterToken(this.ModManifest, "SkinTone", () =>
-            {
-                /*created save
-                if (CreatedSave == true)
-                    return new[] { $"{Game1.MasterPlayer.skinColor}" };
-                if (Game1.MasterPlayer.skin.Value is not 0)
-                    return new[] { $"{Game1.MasterPlayer.skin.Value}" };*/
-
-                if (Game1.player.skin.Value is not 0)
-                    return new[] { $"{Game1.player.skin.Value}" };
-
-                // save is loaded
-                if (Context.IsWorldReady)
-                    return new[] { $"{Game1.player.skin.Value}" };
-
-                // no save loaded (e.g. on the title screen)
-                return new[] { "vanilla" };
-
-                });
-        }
-        private void LoadStageChange(object sender, LoadStageChangedEventArgs e)
+        api.RegisterToken(this.ModManifest, "SkinTone", () =>
         {
-            if (e.NewStage.Equals(LoadStage.CreatedBasicInfo) || e.NewStage.HasFlag(LoadStage.CreatedBasicInfo))
-            {
-                this.Monitor.Log("Created basic info");
-                Helper.GameContent.InvalidateCache("Minigames/jojacorps");
+            /*created save
+            if (CreatedSave == true)
+                return new[] { $"{Game1.MasterPlayer.skinColor}" };
+            if (Game1.MasterPlayer.skin.Value is not 0)
+                return new[] { $"{Game1.MasterPlayer.skin.Value}" };*/
 
-                this.Monitor.Log($"Skin value: {Game1.player.skin.Value}");
-            }
+            if (Game1.player.skin.Value is not 0)
+                return new[] { $"{Game1.player.skin.Value}" };
+
+            // save is loaded
+            if (Context.IsWorldReady)
+                return new[] { $"{Game1.player.skin.Value}" };
+
+            // no save loaded (e.g. on the title screen)
+            return new[] { "vanilla" };
+
+        });
+    }
+    private void LoadStageChange(object sender, LoadStageChangedEventArgs e)
+    {
+        if (e.NewStage.Equals(LoadStage.CreatedBasicInfo) || e.NewStage.HasFlag(LoadStage.CreatedBasicInfo))
+        {
+            this.Monitor.Log("Created basic info");
+            Helper.GameContent.InvalidateCache("Minigames/jojacorps");
+
+            this.Monitor.Log($"Skin value: {Game1.player.skin.Value}");
         }
     }
 }
-
-
